Scripts: Include YAML files in prettification
